+++ src/lib/cmp.test.js
@@ -137,6 +137,9 @@ describe('getTCData', () => {
 						cmpId: 10,
 						cmpVersion: 6,
 						eventStatus: 'useractioncomplete',
+						purpose: {
+							legitimateInterests: {foo: 'bar'}
+						}
 					}, true);
 					break;
 				case 'removeEventListener':
@@ -179,6 +182,9 @@ describe('getTCData', () => {
 						cmpId: 10,
 						cmpVersion: 6,
 						eventStatus: 'tcloaded',
+						purpose: {
+							legitimateInterests: {foo: 'bar'}
+						}
 					}, true);
 					break;
 				case 'removeEventListener':